common makeSelfEdge: take 'cnt' as a 'size_t'
Pays down some cast noise at the cost of introducing some more. These introduced
casts will be removed in a future commit.

+++ lib/common/splines.c
@@ -1180,10 +1180,8 @@ double selfRightSpace(edge_t *e) {
  * FIX: With this bias, labels tend to be placed on top of each other.
  * Perhaps for self-edges, the label should be centered.
  */
-void
-makeSelfEdge(edge_t * edges[], int ind, int cnt, double sizex,
-	     double sizey, splineInfo * sinfo)
-{
+void makeSelfEdge(edge_t *edges[], int ind, size_t cnt, double sizex,
+                  double sizey, splineInfo * sinfo) {
     edge_t *e = edges[ind];
 
     /* self edge without ports or
@@ -1196,7 +1194,7 @@ makeSelfEdge(edge_t * edges[], int ind, int cnt, double sizex,
          !(ED_head_port(e).side & LEFT) &&
           (ED_tail_port(e).side != ED_head_port(e).side ||
           !(ED_tail_port(e).side & (TOP|BOTTOM))))) {
-	selfRight(edges, ind, cnt, sizex, sizey, sinfo);
+	selfRight(edges, ind, (int)cnt, sizex, sizey, sinfo);
     }
 
     /* self edge with port on left side */
@@ -1204,19 +1202,19 @@ makeSelfEdge(edge_t * edges[], int ind, int cnt, double sizex,
 
 	/* handle L-R specially */
 	if ((ED_tail_port(e).side & RIGHT) || (ED_head_port(e).side & RIGHT)) {
-	    selfTop(edges, ind, cnt, sizex, sizey, sinfo);
+	    selfTop(edges, ind, (int)cnt, sizex, sizey, sinfo);
 	}
 	else {
-	    selfLeft(edges, ind, cnt, sizex, sizey, sinfo);
+	    selfLeft(edges, ind, (int)cnt, sizex, sizey, sinfo);
 	}
     }
 
     /* self edge with both ports on top side */
     else if (ED_tail_port(e).side & TOP) {
-	selfTop(edges, ind, cnt, sizex, sizey, sinfo);
+	selfTop(edges, ind, (int)cnt, sizex, sizey, sinfo);
     }
     else if (ED_tail_port(e).side & BOTTOM) {
-	selfBottom(edges, ind, cnt, sizex, sizey, sinfo);
+	selfBottom(edges, ind, (int)cnt, sizex, sizey, sinfo);
     }
 
     else assert(0);
